+++ example/DRPLoadingSpinner/ViewController.m
@@ -26,15 +26,16 @@
     
     self.spinner = [[DRPLoadingSpinner alloc] initWithFrame:CGRectMake(0, 0, 28, 28)];
     self.spinner.center = self.view.center;
-    self.spinner.rotationCycleDuration = 2;
-    self.spinner.minimumArcLength = M_PI / 4;
-    self.spinner.drawCycleDuration = 1;
+    self.spinner.rotationCycleDuration = 4;
+    self.spinner.drawCycleDuration = .5;
     self.spinner.lineWidth = 4;
+    self.spinner.maximumArcLength = M_PI + M_PI_2;
+    self.spinner.minimumArcLength = M_PI;
     [self.view addSubview:self.spinner];
     
     self.spinner2 = [[DRPLoadingSpinner alloc] initWithFrame:CGRectMake(0, 0, 28, 28)];
-    self.spinner2.rotationCycleDuration = 2;
-    self.spinner2.drawCycleDuration = .5;
+    self.spinner2.rotationCycleDuration = 3;
+    self.spinner2.drawCycleDuration = 2;
     [self.view addSubview:self.spinner2];
     
     CGRect spinner2Frame = self.spinner.frame;